Change all probability types from integer to double Remove the scale functions that turned double probabilities into uint32_t

+++ src/packet/l4s_packet_queue.hh
@@ -40,7 +40,7 @@ private:
 public:
     L4SPacketQueue( const std::string & args );
 
-    uint32_t calculate_l4s_native_prob ( uint64_t qdelay );
+    double calculate_l4s_native_prob ( uint64_t qdelay );
 
 };
